ruff clean up

+++ wzk/cpp2py/min_sphere.py
@@ -2,7 +2,7 @@ import numpy as np
 import importlib
 
 try:
-    MinSphere = importlib.import_module('wzk.cpp2py.MinSphere.wzkMinSphere')
+    MinSphere = importlib.import_module("wzk.cpp2py.MinSphere.wzkMinSphere")
 
 except ModuleNotFoundError:
     MinSphere = None
@@ -10,7 +10,7 @@ except ModuleNotFoundError:
 
 def min_sphere(x, r) -> (np.ndarray, float):
     n, d = np.shape(x)
-    res = np.zeros(d + 1, dtype='f4', order='c')
+    res = np.zeros(d + 1, dtype="f4", order="c")
 
     if n == 0:
         return np.zeros(d), 0.0
@@ -18,11 +18,11 @@ def min_sphere(x, r) -> (np.ndarray, float):
     if d < 2 or 4 < d:
         raise ValueError
     if d == 2:
-        MinSphere.min_sphere2(x=x.astype(dtype='f4', order='c'), r=r.astype(dtype='f4', order='c'), n=n, res=res)
+        MinSphere.min_sphere2(x=x.astype(dtype="f4", order="c"), r=r.astype(dtype="f4", order="c"), n=n, res=res)
     elif d == 3:
-        MinSphere.min_sphere3(x=x.astype(dtype='f4', order='c'), r=r.astype(dtype='f4', order='c'), n=n, res=res)
+        MinSphere.min_sphere3(x=x.astype(dtype="f4", order="c"), r=r.astype(dtype="f4", order="c"), n=n, res=res)
     elif d == 4:
-        MinSphere.min_sphere4(x=x.astype(dtype='f4', order='c'), r=r.astype(dtype='f4', order='c'), n=n, res=res)
+        MinSphere.min_sphere4(x=x.astype(dtype="f4", order="c"), r=r.astype(dtype="f4", order="c"), n=n, res=res)
     else:
         raise ValueError
 
@@ -38,10 +38,10 @@ def test_min_sphere():
     x0, r0 = min_sphere(x=x, r=r)
 
     fig, ax = new_fig(aspect=1)
-    ax.plot(*x.T, marker='o', ls='', color='red')
-    ax.plot(*x0, marker='x', ls='', color='blue')
-    plot_circles(ax=ax, x=x0, r=r0, alpha=0.5, color='blue')
+    ax.plot(*x.T, marker="o", ls="", color="red")
+    ax.plot(*x0, marker="x", ls="", color="blue")
+    plot_circles(ax=ax, x=x0, r=r0, alpha=0.5, color="blue")
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     test_min_sphere()
